requirements.txt
colorama
matplotlib
numpy
Pillow
scipy
scikit-image
opencv-python